v0.122.0: A17 god-module decomposition & test coverage closure (H17-02)
Split all 8 source files exceeding 1,000 LOC into smaller sub-modules:
- sparql/expr/functions.rs → 7 sub-modules (string, datetime, numeric, iri,
  aggregate, geo, temporal) — thin dispatch table remains at ~90 LOC
- storage/ops/scan.rs → dedup.rs extracted (deduplicate_all/predicate,
  statement_id_for_triple)
- bulk_load.rs → bulk_load/ directory module with json_ingest.rs and
  confidence.rs sub-modules
- admin_handlers.rs → directory module with explorer.rs and diagnostic.rs
- llm/mod.rs → mapping.rs extracted (suggest_mappings, kge_entity_similarity)
- datalog/compiler/mod.rs → helpers.rs extracted (build_join_cond,
  render_comparison_term, compare_op_sql, arith_op_sql)
- gucs/registration/storage.rs → storage_late.rs extracted (v0.81.0+ GUCs)
- datalog/parser.rs → parser_tests.rs extracted via #[path]

CI gate tightened from 1,800 to 1,000 LOC (zero fiCI gate tightened from 1,800 to 1,000 LOC (zero fiCI gate tightened from 1,800 to 1,000 LOC (zero fiCI gate tightened from 1,800 toad_CI gate tightened from 1,800 to 1,000 LOC (zero fiCI gate tightenedAT-10..14)
- v0120_tenant_quota.sql (QUOTA-01..04)
- v0122_module_splits.sql (SPLI- v0122_module_splits.sql (SPLI- v0122_module_splits.sql (SPLI- v0122schema changes)
Jena CI badge reference added to docs/src/reference/conformance-trends.md
ROADMAP.md: v0.122.0 marked as Released
CHANGELOG.md: v0.122.0 entry added
