+++ pg_ripple_http/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pg_ripple_http"
-version = "0.121.0"
+version = "0.122.0"
 edition = "2024"
 description = "SPARQL 1.1 Protocol HTTP endpoint for pg_ripple — connects PostgreSQL 18 RDF triple store to the web"
 license = "Apache-2.0"
