+++ roadmap/v0.122.0.md
@@ -1,0 +1,95 @@
+# v0.122.0 — A17 God-Module Decomposition & Test Coverage Closure
+
+**Status:** ✅ Released
+
+## Summary
+
+v0.122.0 decomposes all eight remaining god-modules identified in Assessment 17
+(H17-02) and closes the test coverage gap for all v0.119.0–v0.120.0 features.
+
+After all splits, zero source files exceed 1,000 LOC. The CI file-size lint gate
+is tightened from 1,800 to 1,000 lines. Five new pg_regress test suites are added
+covering diagnostic snapshot, read-replica routing, `compat_check()` JSON schema,
+tenant quota, and module-split regression testing.
+
+## Deliverables
+
+### H17-02 / PERF-M-02 — sparql/expr/functions.rs split (was 1,252 LOC)
+
+- [x] `src/sparql/expr/functions.rs` rewritten as thin dispatch table (~90 LOC)
+- [x] `src/sparql/expr/string.rs` — STR, STRLEN, SUBSTR, UCASE, LCASE, CONCAT,
+  REPLACE, ENCODE_FOR_URI, STRLANG, STRDT, STRBEFORE, STRAFTER, MD5, SHA*, UUID
+- [x] `src/sparql/expr/datetime.rs` — NOW, YEAR, MONTH, DAY, HOURS, MINUTES,
+  SECONDS, TIMEZONE, TZ
+- [x] `src/sparql/expr/numeric.rs` — ABS, CEIL, FLOOR, ROUND, RAND
+- [x] `src/sparql/expr/iri.rs` — IRI/URI, BNODE, LANG, DATATYPE, XSD cast dispatch
+- [x] `src/sparql/expr/aggregate.rs` — aggregate function stub
+- [x] `src/sparql/expr/geo.rs` — GeoSPARQL spatial functions
+- [x] `src/sparql/expr/temporal.rs` — Allen's interval relations, fuzzy string match
+- [x] `src/sparql/expr/mod.rs` updated with all sub-module declarations
+
+### H17-02 / PERF-L-01 — storage/ops/scan.rs split (was 1,171 LOC)
+
+- [x] `src/storage/ops/scan.rs` trimmed to ~976 LOC (BGP scan dispatch)
+- [x] `src/storage/ops/dedup.rs` extracted: `deduplicate_predicate()`,
+  `deduplicate_all()`, `statement_id_for_triple()`
+
+### H17-02 — bulk_load.rs split (was 1,173 LOC)
+
+- [x] `src/bulk_load.rs` converted to `src/bulk_load/` directory module
+- [x] `src/bulk_load/json_ingest.rs` — JSON-LD and JSON to N-Triples ingest
+- [x] `src/bulk_load/confidence.rs` — probabilistic triple loading helpers
+
+### PERF-M-03 — admin_handlers.rs split (was 1,168 LOC)
+
+- [x] `pg_ripple_http/src/routing/admin_handlers.rs` converted to directory module
+- [x] `pg_ripple_http/src/routing/admin_handlers/explorer.rs` — visual graph explorer
+- [x] `pg_ripple_http/src/routing/admin_handlers/diagnostic.rs` — diagnostic snapshot
+
+### PERF-L-02 — llm/mod.rs split (was 1,071 LOC)
+
+- [x] `src/llm/mod.rs` trimmed to ~888 LOC
+- [x] `src/llm/mapping.rs` — Automated Ontology Mapping: `suggest_mappings()`,
+  `kge_entity_similarity()`, `parse_embedding_str()`, `jaccard_similarity()`
+
+### H17-02 — datalog/compiler/mod.rs split (was 1,068 LOC)
+
+- [x] `src/datalog/compiler/mod.rs` trimmed to ~982 LOC
+- [x] `src/datalog/compiler/helpers.rs` — SQL helper functions: `build_join_cond()`,
+  `build_not_exists_conds()`, `render_comparison_term()`, `compare_op_sql()`,
+  `arith_op_sql()`
+
+### H17-02 — gucs/registration/storage.rs split (was 1,058 LOC)
+
+- [x] `src/gucs/registration/storage.rs` trimmed to ~665 LOC
+- [x] `src/gucs/registration/storage_late.rs` — v0.81.0+ GUC registrations
+
+### H17-02 — datalog/parser.rs split (was 1,030 LOC)
+
+- [x] `src/datalog/parser.rs` trimmed to ~985 LOC
+- [x] `src/datalog/parser_tests.rs` — parser test module extracted via `#[path]`
+
+### CI gate tightening
+
+- [x] `lint-file-size` step in `.github/workflows/ci.yml` updated from 1,800-line
+  limit to 1,000-line limit
+- [x] Zero source files exceed 1,000 LOC (verified with `find src ... wc -l`)
+
+### Test coverage — v0.120.0 features
+
+- [x] `tests/pg_regress/sql/v0120_diagnostic_snapshot.sql` — DIAG-01 through DIAG-04
+- [x] `tests/pg_regress/sql/v0120_read_replica_routing.sql` — REPLICA-01 through REPLICA-04
+- [x] `tests/pg_regress/sql/v0120_compat_check.sql` — COMPAT-10 through COMPAT-14
+- [x] `tests/pg_regress/sql/v0120_tenant_quota.sql` — QUOTA-01 through QUOTA-04
+- [x] `tests/pg_regress/sql/v0122_module_splits.sql` — SPLIT-01 through SPLIT-05
+
+### Conformance infrastructure
+
+- [x] WatDiv correctness gate: `watdiv-suite` CI job runs with `continue-on-error: false`
+  and validates row counts against baselines (already enforced since v0.91.0)
+- [x] Jena CI badge: badge JSON written to `results/jena-badge.json` by
+  `jena-suite` CI job; pass rate displayed in `docs/src/reference/conformance-trends.md`
+
+### Migration script
+
+- [x] `sql/pg_ripple--0.121.0--0.122.0.sql` created (no schema changes)
