+++ src/bulk_load/mod.rs
@@ -14,6 +14,15 @@
 //! File-path variants read the file content via `pg_read_file()` (superuser-only
 //! PostgreSQL built-in) and then delegate to the inline TEXT variants.
 
+//! v0.122.0 H17-02: JSON ingest helpers extracted to `json_ingest`,
+//! confidence loader extracted to `confidence`.
+
+pub(crate) mod confidence;
+pub(crate) mod json_ingest;
+
+pub use confidence::{ensure_confidence_catalog, load_triples_with_confidence};
+pub use json_ingest::{json_ld_load, json_to_ntriples, json_to_ntriples_and_load};
+
 use std::collections::HashMap;
 use std::sync::atomic::{AtomicU8, Ordering};
 
@@ -684,490 +693,4 @@ fn parse_nt_star_literal(s: &str) -> Option<(i64, &str)> {
         let id = dictionary::encode_plain_literal(&value);
         Some((id, rest))
     }
-}
-
-// ─── JSON → N-Triples (v0.52.0) ──────────────────────────────────────────────
-
-/// Escape a string for safe use inside an N-Triples double-quoted literal.
-fn escape_nt_literal(s: &str) -> String {
-    let mut out = String::with_capacity(s.len() + 4);
-    for c in s.chars() {
-        match c {
-            '"' => out.push_str("\\\""),
-            '\\' => out.push_str("\\\\"),
-            '\n' => out.push_str("\\n"),
-            '\r' => out.push_str("\\r"),
-            '\t' => out.push_str("\\t"),
-            _ => out.push(c),
-        }
-    }
-    out
-}
-
-/// Convert a `serde_json::Value` to an N-Triples object term string.
-///
-/// Objects become a blank node with further triples (emitted into `extra`).
-/// Arrays become repeated triples on the caller side.
-fn json_value_to_nt_term(
-    val: &serde_json::Value,
-    context: &std::collections::HashMap<String, String>,
-    bn_counter: &mut u64,
-    extra: &mut String,
-) -> Option<String> {
-    match val {
-        serde_json::Value::String(s) => Some(format!("\"{}\"", escape_nt_literal(s))),
-        serde_json::Value::Number(n) => {
-            // RT-FIX-06: check is_f64() first so that `5.0` is stored as
-            // xsd:decimal rather than collapsing to xsd:integer.
-            if n.is_f64() {
-                n.as_f64()
-                    .map(|f| format!("\"{}\"^^<http://www.w3.org/2001/XMLSchema#decimal>", f))
-            } else if let Some(i) = n.as_i64() {
-                Some(format!(
-                    "\"{}\"^^<http://www.w3.org/2001/XMLSchema#integer>",
-                    i
-                ))
-            } else if let Some(u) = n.as_u64() {
-                // Values in (i64::MAX, u64::MAX] — still valid xsd:integer.
-                Some(format!(
-                    "\"{}\"^^<http://www.w3.org/2001/XMLSchema#integer>",
-                    u
-                ))
-            } else {
-                // RT-FIX-04B: numbers that exceed u64::MAX are preserved as
-                // an xsd:integer string rather than silently losing precision.
-                let s = n.to_string();
-                if s.contains('.') || s.contains('e') || s.contains('E') {
-                    Some(format!(
-                        "\"{}\"^^<http://www.w3.org/2001/XMLSchema#decimal>",
-                        s
-                    ))
-                } else {
-                    Some(format!(
-                        "\"{}\"^^<http://www.w3.org/2001/XMLSchema#integer>",
-                        s
-                    ))
-                }
-            }
-        }
-        serde_json::Value::Bool(b) => Some(format!(
-            "\"{}\"^^<http://www.w3.org/2001/XMLSchema#boolean>",
-            b
-        )),
-        serde_json::Value::Null => None,
-        serde_json::Value::Object(map) => {
-            *bn_counter += 1;
-            let bn = format!("_:b{}", bn_counter);
-            for (k, v) in map {
-                let pred_iri = resolve_key_to_iri(k, context);
-                json_object_to_ntriples_inner(&bn, &pred_iri, v, context, bn_counter, extra);
-            }
-            Some(bn)
-        }
-        serde_json::Value::Array(_) => None, // arrays handled at the caller level
-    }
-}
-
-/// Resolve a JSON key to a full IRI using the provided context mapping.
-///
-/// The context maps short keys (e.g. `"name"`) to full IRIs
-/// (e.g. `"https://schema.org/name"`).  If no mapping exists the key is used
-/// as-is, which produces a relative IRI (acceptable for local development).
-fn resolve_key_to_iri(key: &str, context: &std::collections::HashMap<String, String>) -> String {
-    context.get(key).cloned().unwrap_or_else(|| key.to_owned())
-}
-
-/// Recursively emit N-Triples for one `(subject, predicate, value)` combination.
-fn json_object_to_ntriples_inner(
-    subject: &str,
-    pred_iri: &str,
-    val: &serde_json::Value,
-    context: &std::collections::HashMap<String, String>,
-    bn_counter: &mut u64,
-    out: &mut String,
-) {
-    let pred_term = format!("<{}>", pred_iri);
-    match val {
-        serde_json::Value::Array(items) => {
-            for item in items {
-                json_object_to_ntriples_inner(subject, pred_iri, item, context, bn_counter, out);
-            }
-        }
-        _ => {
-            if let Some(obj_term) = json_value_to_nt_term(val, context, bn_counter, out) {
-                let s_term = if subject.starts_with("_:") {
-                    subject.to_owned()
-                } else {
-                    format!("<{}>", subject)
-                };
-                out.push_str(&s_term);
-                out.push(' ');
-                out.push_str(&pred_term);
-                out.push(' ');
-                out.push_str(&obj_term);
-                out.push_str(" .\n");
-            }
-        }
-    }
-}
-
-/// Convert a flat JSON object to N-Triples, mapping each key to a predicate IRI.
-///
-/// - `payload` — JSON object; keys become predicates, values become objects.
-/// - `subject_iri` — IRI for the RDF subject (without angle brackets).
-/// - `type_iri` — optional `rdf:type` IRI.  When provided, a `<subject>
-///   <rdf:type> <type_iri> .` triple is prepended.
-/// - `context` — optional JSONB `{"key": "iri", …}` mapping.  When `None`,
-///   keys are used verbatim as IRIs (useful when keys are already full IRIs or
-///   the caller applies vocabulary alignment via Datalog rules separately).
-///
-/// Returns the N-Triples string.  Nested objects become blank nodes.  Arrays
-/// produce one triple per element.  `null` values are silently skipped.
-///
-/// RT-FIX-07: Validate that a JSON key is safe to expand under @vocab.
-///
-/// Raises a PostgreSQL error if the key contains characters forbidden in IRI
-/// references (RFC 3987): spaces, control characters, `"`, `<`, `>`, `{`, `}`,
-/// `|`, `\\`, `^`, and `` ` ``.
-fn validate_iri_key_or_error(key: &str) {
-    let forbidden =
-        |c: char| c <= '\x20' || matches!(c, '"' | '<' | '>' | '{' | '}' | '|' | '\\' | '^' | '`');
-    if let Some(bad) = key.chars().find(|&c| forbidden(c)) {
-        pgrx::error!(
-            "cannot derive predicate IRI from JSON key {:?}: \
-             character {:?} is not allowed in IRI references — \
-             add an explicit context entry, e.g. {:?}: \"ex:{}\"",
-            key,
-            bad,
-            key,
-            key.replace(' ', "_")
-        );
-    }
-}
-
-pub fn json_to_ntriples(
-    payload: &serde_json::Value,
-    subject_iri: &str,
-    type_iri: Option<&str>,
-    context_map: Option<&serde_json::Value>,
-) -> String {
-    let map = match payload {
-        serde_json::Value::Object(m) => m,
-        _ => {
-            pgrx::warning!("json_to_ntriples: payload must be a JSON object");
-            return String::new();
-        }
-    };
-
-    // Build context: key → IRI lookup table.
-    let mut ctx: std::collections::HashMap<String, String> = std::collections::HashMap::new();
-    if let Some(serde_json::Value::Object(c)) = context_map {
-        if let Some(serde_json::Value::String(vocab)) = c.get("@vocab") {
-            // @vocab provides the default IRI prefix for all unmapped keys.
-            for (k, _) in map {
-                if !k.starts_with('@') && !c.contains_key(k.as_str()) {
-                    // RT-FIX-07: validate the key is a valid IRI local part.
-                    validate_iri_key_or_error(k);
-                    ctx.insert(k.clone(), format!("{}{}", vocab, k));
-                }
-            }
-        }
-        for (k, v) in c {
-            if k == "@vocab" || k.starts_with('@') {
-                continue;
-            }
-            match v {
-                serde_json::Value::String(iri) => {
-                    // Simple string-valued context entry: "key": "iri"
-                    ctx.insert(k.clone(), iri.clone());
-                }
-                serde_json::Value::Object(meta) => {
-                    // BUG-JSONLD-CONTEXT-01: object-form context entry.
-                    // Extract the @id field as the predicate IRI.
-                    if let Some(serde_json::Value::String(iri)) = meta.get("@id") {
-                        ctx.insert(k.clone(), iri.clone());
-                    } else if let Some(serde_json::Value::String(vocab)) = c.get("@vocab") {
-                        // Fall back to @vocab expansion for the key.
-                        ctx.insert(k.clone(), format!("{}{}", vocab, k));
-                    }
-                    // Note: @container, @type, @language metadata is noted
-                    // for forward-compatibility but not fully processed here;
-                    // the IRI mapping is the critical fix for round-trip correctness.
-                }
-                _ => {}
-            }
-        }
-    }
-
-    let mut out = String::with_capacity(512);
-    let mut bn_counter: u64 = 0;
-
-    // Emit rdf:type triple if requested.
-    if let Some(t) = type_iri {
-        out.push_str(&format!(
-            "<{}> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <{}> .\n",
-            subject_iri, t
-        ));
-    }
-
-    // Emit one triple per key/value pair in the payload.
-    for (key, val) in map {
-        if key.starts_with('@') {
-            continue; // skip JSON-LD keywords
-        }
-        let pred_iri = resolve_key_to_iri(key, &ctx);
-        json_object_to_ntriples_inner(subject_iri, &pred_iri, val, &ctx, &mut bn_counter, &mut out);
-    }
-
-    out
-}
-
-/// Load the N-Triples produced by `json_to_ntriples` directly into the store.
-///
-/// This is the common path used by the `json_to_ntriples_and_load` SQL function:
-/// it converts the JSON to N-Triples then immediately loads them, returning the
-/// count of triples inserted.
-pub fn json_to_ntriples_and_load(
-    payload: &serde_json::Value,
-    subject_iri: &str,
-    type_iri: Option<&str>,
-    context_map: Option<&serde_json::Value>,
-) -> i64 {
-    let nt = json_to_ntriples(payload, subject_iri, type_iri, context_map);
-    if nt.is_empty() {
-        return 0;
-    }
-    load_ntriples(&nt, false)
-}
-
-// ─── JSONLD-INGEST-02: multi-subject JSON-LD document ingest ─────────────────
-
-/// Ingest a full JSON-LD document that may contain multiple top-level subjects.
-///
-/// Handles both the `@graph` form (multiple top-level nodes) and the single-node
-/// form (object with `@id`).  Each top-level node must have an `@id` key.
-///
-/// - `document`      — JSONB value representing the JSON-LD document.
-/// - `default_graph` — named graph IRI to use when the document has no outer
-///   named graph.  `None` means the default graph (id = 0).
-///
-/// Returns the total number of triples loaded.
-pub fn json_ld_load(document: &serde_json::Value, default_graph: Option<&str>) -> i64 {
-    // Determine the outer named graph (if the document has a top-level @id
-    // that looks like a graph IRI, treat it as the target graph).
-    let outer_graph: Option<String> = match document {
-        serde_json::Value::Object(obj) => obj
-            .get("@id")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_owned()),
-        _ => None,
-    };
-
-    // Collect the nodes to process.
-    let nodes: Vec<&serde_json::Value> = match document {
-        serde_json::Value::Object(obj) => {
-            if let Some(serde_json::Value::Array(graph)) = obj.get("@graph") {
-                // Multi-subject @graph form.
-                graph.iter().collect()
-            } else {
-                // Single-node form.
-                vec![document]
-            }
-        }
-        serde_json::Value::Array(arr) => {
-            // Top-level array (expanded JSON-LD).
-            arr.iter().collect()
-        }
-        _ => {
-            pgrx::error!("json_ld_load: document must be a JSON object or array");
-        }
-    };
-
-    let mut total = 0i64;
-
-    for node in nodes {
-        let obj = match node {
-            serde_json::Value::Object(o) => o,
-            _ => {
-                pgrx::warning!("json_ld_load: skipping non-object node in @graph");
-                continue;
-            }
-        };
-
-        // Each node must have an @id.
-        let subject_iri = obj
-            .get("@id")
-            .and_then(|v| v.as_str())
-            .unwrap_or_else(|| {
-                pgrx::error!(
-                    "json_ld_load: top-level node in @graph is missing @id; \
-                     all JSON-LD nodes must have an @id when using json_ld_load(). \
-                     Provide an @id field or use json_to_ntriples_and_load() with an explicit subject IRI."
-                )
-            });
-
-        // Build context from the node's @context (if present) or the outer document's @context.
-        let ctx = node
-            .get("@context")
-            .or_else(|| document.as_object().and_then(|d| d.get("@context")));
-
-        // Determine the target graph.
-        let graph_id: i64 = {
-            let g_iri = outer_graph
-                .as_deref()
-                .or(default_graph)
-                .filter(|s| !s.is_empty());
-            match g_iri {
-                None => 0, // default graph
-                Some(g) => dictionary::encode(g, dictionary::KIND_IRI),
-            }
-        };
-
-        // Build a payload without @-keywords for the json_to_ntriples path.
-        let payload_obj: serde_json::Map<String, serde_json::Value> = obj
-            .iter()
-            .filter(|(k, _)| !k.starts_with('@'))
-            .map(|(k, v)| (k.clone(), v.clone()))
-            .collect();
-
-        if payload_obj.is_empty() {
-            continue;
-        }
-
-        let payload = serde_json::Value::Object(payload_obj);
-        let nt = json_to_ntriples(&payload, subject_iri, None, ctx);
-        if nt.is_empty() {
-            continue;
-        }
-
-        total += if graph_id == 0 {
-            load_ntriples(&nt, false)
-        } else {
-            load_ntriples_into_graph(&nt, graph_id)
-        };
-    }
-
-    total
-}
-
-// ─── v0.87.0 LOAD-CONF-01: confidence-aware bulk loader ──────────────────────
-
-/// Ensure `_pg_ripple.confidence` table and its index exist.
-///
-/// This is idempotent and safe to call on every `load_triples_with_confidence`
-/// and `vacuum_confidence` invocation.  On a fresh install (before the v0.87.0
-/// migration script has been applied), `CREATE TABLE IF NOT EXISTS` creates the
-/// table on-demand; on upgraded instances it is a no-op.
-pub(crate) fn ensure_confidence_catalog() {
-    pgrx::Spi::run(
-        "CREATE TABLE IF NOT EXISTS _pg_ripple.confidence ( \
-            statement_id BIGINT  NOT NULL, \
-            confidence   FLOAT8  NOT NULL CHECK (confidence >= 0.0 AND confidence <= 1.0), \
-            model        TEXT    NOT NULL DEFAULT 'datalog', \
-            asserted_at  TIMESTAMPTZ NOT NULL DEFAULT now(), \
-            PRIMARY KEY  (statement_id, model) \
-        )",
-    )
-    .unwrap_or_else(|e| pgrx::warning!("confidence catalog creation: {e}"));
-
-    pgrx::Spi::run(
-        "CREATE INDEX IF NOT EXISTS confidence_stmt_idx \
-         ON _pg_ripple.confidence (statement_id)",
-    )
-    .unwrap_or_else(|e| pgrx::warning!("confidence_stmt_idx creation: {e}"));
-}
-
-/// Load triples with an explicit uniform confidence score.
-///
-/// After inserting triples, inserts confidence rows with the given confidence
-/// score for every newly-inserted SID into `_pg_ripple.confidence (model='explicit')`.
-///
-/// `format` may be `'ntriples'` (default), `'nquads'`, `'turtle'`, or `'jsonld'`.
-/// `graph_uri` routes all triples to a named graph when provided.
-///
-/// Returns the number of triples loaded.
-pub fn load_triples_with_confidence(
-    data: &str,
-    confidence: f64,
-    format: &str,
-    graph_uri: Option<&str>,
-) -> i64 {
-    // Validate confidence value: reject NaN, Inf, and out-of-range values.
-    // M15-09 (v0.95.0): explicit NaN and Inf checks come first so that the
-    // error message identifies the specific problem rather than just saying
-    // "outside [0.0, 1.0]".
-    if confidence.is_nan() {
-        pgrx::error!("confidence value is NaN — must be a finite number in [0.0, 1.0] (PT0301)");
-    }
-    if confidence.is_infinite() {
-        pgrx::error!(
-            "confidence value is {} — must be a finite number in [0.0, 1.0] (PT0301)",
-            if confidence.is_sign_positive() {
-                "+Infinity"
-            } else {
-                "-Infinity"
-            }
-        );
-    }
-    if !(0.0..=1.0).contains(&confidence) {
-        pgrx::error!(
-            "confidence must be in [0.0, 1.0]; got {} (PT0301)",
-            confidence
-        );
-    }
-
-    // Ensure the confidence catalog table exists (idempotent; works on fresh installs).
-    ensure_confidence_catalog();
-
-    // Load the triples using the appropriate loader.
-    let count = match format.to_ascii_lowercase().as_str() {
-        "ntriples" | "nt" => {
-            if let Some(g_uri) = graph_uri {
-                let g_id = crate::dictionary::encode(g_uri, crate::dictionary::KIND_IRI);
-                load_ntriples_into_graph(data, g_id)
-            } else {
-                load_ntriples(data, false)
-            }
-        }
-        "nquads" | "nq" => load_nquads(data, false),
-        "turtle" | "ttl" => {
-            if let Some(g_uri) = graph_uri {
-                let g_id = crate::dictionary::encode(g_uri, crate::dictionary::KIND_IRI);
-                load_turtle_into_graph(data, g_id)
-            } else {
-                load_turtle(data, false)
-            }
-        }
-        other => {
-            pgrx::error!(
-                "unsupported format '{}'; use 'ntriples', 'nquads', 'turtle'",
-                other
-            );
-        }
-    };
-
-    if count > 0 {
-        // Insert confidence rows for all SIDs inserted in the current transaction.
-        // We use vp_rare and dedicated VP tables via the predicates catalog.
-        // For simplicity, we tag the most recently inserted SIDs across all VP tables.
-        let conf_sql = format!(
-            "INSERT INTO _pg_ripple.confidence (statement_id, confidence, model) \
-             SELECT i, {confidence}::float8, 'explicit' \
-             FROM _pg_ripple.vp_rare \
-             WHERE source = 0 \
-             ORDER BY i DESC LIMIT {count} \
-             ON CONFLICT (statement_id, model) DO NOTHING"
-        );
-        if let Err(e) = pgrx::Spi::run_with_args(&conf_sql, &[]) {
-            pgrx::warning!("confidence insert error in load_triples_with_confidence: {e}");
-        }
-
-        // PERF-06 (v0.90.0): run ANALYZE so the planner has fresh statistics
-        // on the confidence table for subsequent confidence-join queries.
-        if let Err(e) = pgrx::Spi::run("ANALYZE _pg_ripple.confidence") {
-            pgrx::warning!("load_triples_with_confidence: ANALYZE failed: {e}");
-        }
-    }
-
-    count
 }
